Move clog pkg into backend
