+++ backend/app.go
@@ -9,8 +9,8 @@ import (
 	"time"
 
 	"cloud.google.com/go/firestore"
+	"github.com/nownabe/golink/backend/clog"
 	"github.com/nownabe/golink/backend/middleware"
-	"github.com/nownabe/golink/go/clog"
 	"github.com/nownabe/golink/go/errors"
 	"golang.org/x/net/http2"
 	"golang.org/x/net/http2/h2c"
