+++ backend/interceptor/interceptors.go
@@ -5,7 +5,7 @@ import (
 	"net/http"
 
 	"github.com/bufbuild/connect-go"
-	"github.com/nownabe/golink/go/clog"
+	"github.com/nownabe/golink/backend/clog"
 	"github.com/nownabe/golink/go/errors"
 )
