+++ backend/middleware/recoverer.go
@@ -3,7 +3,7 @@ package middleware
 import (
 	"net/http"
 
-	"github.com/nownabe/golink/go/clog"
+	"github.com/nownabe/golink/backend/clog"
 	"github.com/nownabe/golink/go/errors"
 )
