+++ backend/redirect_handler.go
@@ -9,7 +9,7 @@ import (
 	"time"
 
 	"cloud.google.com/go/firestore"
-	"github.com/nownabe/golink/go/clog"
+	"github.com/nownabe/golink/backend/clog"
 	"github.com/nownabe/golink/go/errors"
 )
