+++ go/errors/errors.go
@@ -5,8 +5,6 @@ import (
 	"errors"
 	"fmt"
 	"runtime"
-
-	"github.com/nownabe/golink/go/clog"
 )
 
 func New(msg string) error {
@@ -64,10 +62,10 @@ func Is(err error, target error) bool {
 }
 
 type wrapped struct {
-	err     error
-	msg     string
-	stack   []uintptr
-	context *clog.ErrorContext
+	err   error
+	msg   string
+	stack []uintptr
+	// TODO: context *clog.ErrorContext
 }
 
 func (w *wrapped) Error() string {
@@ -81,9 +79,11 @@ func (w *wrapped) Unwrap() error {
 	return w.err
 }
 
+/*
 func (w *wrapped) ErrorContext() *clog.ErrorContext {
 	return w.context
 }
+*/
 
 func (w *wrapped) Stack() []byte {
 	buf := bytes.Buffer{}
